[extension-selenium] Make context coordinates/size variables available for mobile app profile
