+++ vividus-extension-selenium/src/main/java/org/vividus/bdd/variable/ui/AbstractContextProvidingDynamicVariable.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2019-2020 the original author or authors.
+ * Copyright 2019-2021 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.vividus.bdd.variable.ui.web;
+package org.vividus.bdd.variable.ui;
 
 import java.util.function.Function;
