+++ vividus-extension-selenium/src/main/resources/spring.xml
@@ -12,6 +12,7 @@
     <context:component-scan base-package="org.vividus.bdd.converter.ui" >
          <context:exclude-filter type="regex" expression="org\.vividus\.bdd\.converter\.ui\..*\..*"/>
     </context:component-scan>
+    <context:component-scan base-package="org.vividus.bdd.variable.ui"/>
 
     <bean id="parentWebDriverFactory" class="org.vividus.selenium.AbstractWebDriverFactory" abstract="true">
         <property name="remoteDriverUrl" value="${selenium.grid.url}" />
